- move from pip to uv for package management - creating superuser for db is now moved to init_db (single file) - seed_data now uses superuser (no need to register) - some other claude suggested fixes
