+++ backend/.env.example
@@ -1,13 +1,13 @@
 DATABASE_URL=postgresql://postgres:postgres@db:5432/budget_analyzer
 REDIS_URL=redis://redis:6379/0
-SECRET_KEY=your-secret-key-change-this-in-production
+SECRET_KEY=                          # required: generate with `openssl rand -hex 32`
 ALGORITHM=HS256
 ACCESS_TOKEN_EXPIRE_MINUTES=30
-PASSWORD_RESET_TOKEN_EXPIRE_MINUTES=30  # Token expiration
-FIRST_SUPERUSER_EMAIL=admin@example.com
-FIRST_SUPERUSER_PASSWORD=admin123
-FIRST_SUPERUSER_USERNAME=admin
-FIRST_SUPERUSER_FULL_NAME=Administrator
+PASSWORD_RESET_TOKEN_EXPIRE_MINUTES=30
+FIRST_SUPERUSER_EMAIL=              # required
+FIRST_SUPERUSER_PASSWORD=           # required
+FIRST_SUPERUSER_USERNAME=           # required
+FIRST_SUPERUSER_FULL_NAME=
 FRONTEND_URL=http://localhost:5173
 PASSWORD_RESET_PATH=/reset-password
