+++ backend/app/db/session.py
@@ -1,5 +1,8 @@
+from collections.abc import Iterator
+from contextlib import contextmanager
+
 from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
+from sqlalchemy.orm import Session, sessionmaker
 
 from app.core.config import settings
 from app.core.logger_init import setup_logging
@@ -13,3 +16,12 @@ def get_db():
     db = SessionLocal()
     with db as session:
         yield session
+
+
+@contextmanager
+def get_session() -> Iterator[Session]:
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
