+++ docker/backend.Dockerfile
@@ -1,21 +1,18 @@
 FROM python:3.12-slim
+COPY --from=ghcr.io/astral-sh/uv:latest /uv /uvx /bin/
 
 WORKDIR /app
 
 # Install system dependencies
 RUN apt-get update && apt-get install -y \
-    gcc \
     postgresql-client \
     && rm -rf /var/lib/apt/lists/*
 
-# Copy requirements file
-COPY backend/requirements.txt .
-
-# Install Python dependencies
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Copy the application code
-COPY backend/ ./
+# Install dependencies into /opt/venv so the ./backend:/app volume mount doesn't wipe them
+COPY pyproject.toml uv.lock ./
+ENV UV_PROJECT_ENVIRONMENT=/opt/venv
+RUN uv sync --frozen --no-dev
+ENV PATH="/opt/venv/bin:$PATH"
 
 # Expose the FastAPI port
 EXPOSE 8000
