+++ docker/frontend.Dockerfile
@@ -8,9 +8,6 @@ COPY frontend/package*.json ./
 # Install dependencies
 RUN npm install
 
-# Copy the rest of the application
-COPY frontend/ ./
-
 # Expose the Vite dev server port
 EXPOSE 5173
